+++ src/Atoms/Icon/action/arrowNegative/index.ts
@@ -1,0 +1,7 @@
+import ArrowNegative16 from './ArrowNegative16';
+import ArrowNegative24 from './ArrowNegative24';
+
+export default {
+  ArrowNegative16,
+  ArrowNegative24,
+};