+++ src/Atoms/Illustration/index.ts
@@ -35,9 +35,11 @@ import MoneyDeposit64 from './Medium/MoneyDeposit64';
 import MoneyWithdraw64 from './Medium/MoneyWithdraw64';
 import MonthlySavings64 from './Medium/MonthlySavings64';
 import News64 from './Medium/News64';
+import Pension64 from './Medium/Pension64';
 import RiskHigh64 from './Medium/RiskHigh64';
 import RiskLow64 from './Medium/RiskLow64';
 import RiskMedium64 from './Medium/RiskMedium64';
+import Robot64 from './Medium/Robot64';
 import Security64 from './Medium/Security64';
 import StatusError64 from './Medium/StatusError64';
 import StatusFail64 from './Medium/StatusFail64';
@@ -87,11 +89,13 @@ export default {
   News64,
   Orders48,
   Outbox48,
+  Pension64,
   PriceAlert48,
   Refresh48,
   RiskHigh64,
   RiskLow64,
   RiskMedium64,
+  Robot64,
   Search48,
   Security64,
   StatusError64,