+++ src/Molecules/PromotionBanner/PromotionBanner.types.ts
@@ -1,0 +1,21 @@
+import { ColorFn } from 'common/Types';
+import React from 'react';
+
+export type PromotionBannerProps = {
+  /** @default blue */
+  backgroundColor?: 'blue' | 'green' | 'white' | ColorFn;
+  badgeBackground?: ColorFn;
+  badgeContent?: React.ReactNode;
+  children?: React.ReactNode;
+  description?: string | React.ReactChild | React.ReactChild[];
+  /** @default true */
+  dismissible?: boolean;
+  /** @default null */
+  onClose?: () => void;
+  mobileBadgeContent?: React.ReactNode;
+  /** @default module */
+  scope?: 'module' | 'page';
+  title?: string | React.ReactChild | React.ReactChild[];
+};
+
+export type PromotionBannerComponent = React.FC<PromotionBannerProps>;